+++ src/app/containers/FrontPage/index.stories.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import igboData from '#data/igbo/frontpage';
 import pidginData from '#data/pidgin/frontpage';
-import addIdsToItems from '#lib/utilities/preprocessor/rules/addIdsToItems';
 import thaiData from '#data/thai/frontpage';
 import yorubaData from '#data/yoruba/frontpage';
 import punjabiData from '#data/punjabi/frontpage';
-import filterUnknownContentTypes from '#lib/utilities/preprocessor/rules/filterContentType';
-import filterEmptyGroupItems from '#lib/utilities/preprocessor/rules/filterEmptyGroupItems';
-import applySquashTopstories from '#lib/utilities/preprocessor/rules/topstories';
+import addIdsToItems from '#lib/utilities/preprocessor/rules/frontPage/addIdsToItems';
+import filterUnknownContentTypes from '#lib/utilities/preprocessor/rules/frontPage/filterContentType';
+import filterEmptyGroupItems from '#lib/utilities/preprocessor/rules/frontPage/filterEmptyGroupItems';
+import applySquashTopstories from '#lib/utilities/preprocessor/rules/frontPage/topstories';
 import preprocess from '#lib/utilities/preprocessor';
 import FrontPage from '.';
