+++ src/app/lib/utilities/preprocessor/rules/frontPage/filterContentType/index.js
