+++ src/app/lib/utilities/preprocessor/rules/frontPage/filterEmptyGroupItems/index.js
