+++ src/app/lib/utilities/preprocessor/rules/frontPage/filterEmptyGroupItems/index.test.js
