+++ src/app/lib/utilities/preprocessor/rules/frontPage/filterGroupsWithoutStraplines/index.js
