+++ src/app/routes/getInitialData/article/index.js
@@ -2,9 +2,9 @@ import onClient from '#lib/utilities/onClient';
 import getBaseUrl from '../utils/getBaseUrl';
 import fetchData from '../utils/fetchData';
 import { variantSanitiser } from '#lib/utilities/variantHandler';
-import applyTimestampRules from '#lib/utilities/preprocessor/rules/timestamp';
-import addIdsToBlocks from '#lib/utilities/preprocessor/rules/addIdsToBlocks';
-import applyBlockPositioning from '#lib/utilities/preprocessor/rules/blockPositioning';
+import applyTimestampRules from '#lib/utilities/preprocessor/rules/optimoArticle/timestamp';
+import addIdsToBlocks from '#lib/utilities/preprocessor/rules/optimoArticle/addIdsToBlocks';
+import applyBlockPositioning from '#lib/utilities/preprocessor/rules/optimoArticle/blockPositioning';
 
 const getArticleInitialData = async ({ id, service, variant }) => {
   const baseUrl = onClient()
