+++ src/app/routes/getInitialData/frontpage/index.js
@@ -2,11 +2,11 @@ import onClient from '#lib/utilities/onClient';
 import { variantSanitiser } from '#lib/utilities/variantHandler';
 import getBaseUrl from '../utils/getBaseUrl';
 import fetchData from '../utils/fetchData';
-import filterUnknownContentTypes from '#lib/utilities/preprocessor/rules/filterContentType';
-import filterEmptyGroupItems from '#lib/utilities/preprocessor/rules/filterEmptyGroupItems';
-import applySquashTopstories from '#lib/utilities/preprocessor/rules/topstories';
-import addIdsToItems from '#lib/utilities/preprocessor/rules/addIdsToItems';
-import filterGroupsWithoutStraplines from '#lib/utilities/preprocessor/rules/filterGroupsWithoutStraplines';
+import filterUnknownContentTypes from '#lib/utilities/preprocessor/rules/frontPage/filterContentType';
+import filterEmptyGroupItems from '#lib/utilities/preprocessor/rules/frontPage/filterEmptyGroupItems';
+import applySquashTopstories from '#lib/utilities/preprocessor/rules/frontPage/topstories';
+import addIdsToItems from '#lib/utilities/preprocessor/rules/frontPage/addIdsToItems';
+import filterGroupsWithoutStraplines from '#lib/utilities/preprocessor/rules/frontPage/filterGroupsWithoutStraplines';
 
 const preprocessorRules = [
   filterUnknownContentTypes,
